Split updater artifact lifecycle
